Refactor evm-tracing related code for cli
