+++ bin/collator/src/cli.rs
@@ -19,6 +19,9 @@
 use clap::Parser;
 use std::path::PathBuf;
 
+#[cfg(feature = "evm-tracing")]
+use crate::evm_tracing_types::EthApiOptions;
+
 /// An overarching CLI command definition.
 #[derive(Debug, clap::Parser)]
 pub struct Cli {
@@ -30,54 +33,10 @@ pub struct Cli {
     #[clap(flatten)]
     pub run: cumulus_client_cli::RunCmd,
 
-    /// Enable EVM tracing module on a non-authority node.
+    #[allow(missing_docs)]
     #[cfg(feature = "evm-tracing")]
-    #[clap(
-        long,
-        conflicts_with = "collator",
-        conflicts_with = "validator",
-        value_delimiter = ','
-    )]
-    pub ethapi: Vec<EthApi>,
-
-    /// Number of concurrent tracing tasks. Meant to be shared by both "debug" and "trace" modules.
-    #[cfg(feature = "evm-tracing")]
-    #[clap(long, default_value = "10")]
-    pub ethapi_max_permits: u32,
-
-    /// Maximum number of trace entries a single request of `trace_filter` is allowed to return.
-    /// A request asking for more or an unbounded one going over this limit will both return an
-    /// error.
-    #[cfg(feature = "evm-tracing")]
-    #[clap(long, default_value = "500")]
-    pub ethapi_trace_max_count: u32,
-
-    /// Duration (in seconds) after which the cache of `trace_filter` for a given block will be
-    /// discarded.
-    #[cfg(feature = "evm-tracing")]
-    #[clap(long, default_value = "300")]
-    pub ethapi_trace_cache_duration: u64,
-
-    /// Size in bytes of the LRU cache for block data.
-    #[cfg(feature = "evm-tracing")]
-    #[clap(long, default_value = "300000000")]
-    pub eth_log_block_cache: usize,
-
-    /// Size in bytes of the LRU cache for transactions statuses data.
-    #[cfg(feature = "evm-tracing")]
-    #[clap(long, default_value = "300000000")]
-    pub eth_statuses_cache: usize,
-
-    /// Size in bytes of data a raw tracing request is allowed to use.
-    /// Bound the size of memory, stack and storage data.
-    #[cfg(feature = "evm-tracing")]
-    #[clap(long, default_value = "20000000")]
-    pub tracing_raw_max_memory_usage: usize,
-
-    /// Maximum number of logs in a query.
-    #[cfg(feature = "evm-tracing")]
-    #[clap(long, default_value = "10000")]
-    pub max_past_logs: u32,
+    #[clap(flatten)]
+    pub eth_api_options: EthApiOptions,
 
     /// Enable Ethereum compatible JSON-RPC servers (disabled by default).
     #[clap(name = "enable-evm-rpc", long)]
@@ -95,6 +54,7 @@ pub struct Cli {
     #[clap(long, default_value = "50")]
     pub proposer_soft_deadline_percent: u8,
 }
+
 
 /// Possible subcommands of the main binary.
 #[derive(Debug, clap::Subcommand)]
@@ -183,58 +143,3 @@ impl RelayChainCli {
     }
 }
 
-/// EVM tracing CLI flags.
-#[cfg(feature = "evm-tracing")]
-#[derive(Debug, PartialEq, Clone)]
-pub enum EthApi {
-    /// Enable EVM debug RPC methods.
-    Debug,
-    /// Enable EVM trace RPC methods.
-    Trace,
-    /// Enable pending transactions RPC methods.
-    TxPool,
-}
-
-#[cfg(feature = "evm-tracing")]
-impl std::str::FromStr for EthApi {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "debug" => Self::Debug,
-            "trace" => Self::Trace,
-            "txpool" => Self::TxPool,
-            _ => {
-                return Err(format!(
-                    "`{}` is not recognized as a supported Ethereum Api",
-                    s
-                ))
-            }
-        })
-    }
-}
-
-/// EVM tracing CLI config.
-#[cfg(feature = "evm-tracing")]
-pub struct EvmTracingConfig {
-    /// Enabled EVM tracing flags.
-    pub ethapi: Vec<EthApi>,
-    /// Number of concurrent tracing tasks.
-    pub ethapi_max_permits: u32,
-    /// Maximum number of trace entries a single request of `trace_filter` is allowed to return.
-    /// A request asking for more or an unbounded one going over this limit will both return an
-    /// error.
-    pub ethapi_trace_max_count: u32,
-    /// Duration (in seconds) after which the cache of `trace_filter` for a given block will be
-    /// discarded.
-    pub ethapi_trace_cache_duration: u64,
-    /// Size in bytes of the LRU cache for block data.
-    pub eth_log_block_cache: usize,
-    /// Size in bytes of the LRU cache for transactions statuses data.
-    pub eth_statuses_cache: usize,
-    /// Maximum number of logs in a query.
-    pub max_past_logs: u32,
-    /// Size in bytes of data a raw tracing request is allowed to use.
-    /// Bound the size of memory, stack and storage data.
-    pub tracing_raw_max_memory_usage: usize,
-}
