+++ bin/collator/src/command.rs
@@ -847,15 +847,15 @@ pub fn run() -> Result<()> {
             let collator_options = cli.run.collator_options();
 
             #[cfg(feature = "evm-tracing")]
-            let evm_tracing_config = crate::cli::EvmTracingConfig {
-                ethapi: cli.ethapi,
-                ethapi_max_permits: cli.ethapi_max_permits,
-                ethapi_trace_max_count: cli.ethapi_trace_max_count,
-                ethapi_trace_cache_duration: cli.ethapi_trace_cache_duration,
-                eth_log_block_cache: cli.eth_log_block_cache,
-                eth_statuses_cache: cli.eth_statuses_cache,
-                max_past_logs: cli.max_past_logs,
-                tracing_raw_max_memory_usage: cli.tracing_raw_max_memory_usage,
+            let evm_tracing_config = crate::evm_tracing_types::EvmTracingConfig {
+                ethapi: cli.eth_api_options.ethapi,
+                ethapi_max_permits: cli.eth_api_options.ethapi_max_permits,
+                ethapi_trace_max_count: cli.eth_api_options.ethapi_trace_max_count,
+                ethapi_trace_cache_duration: cli.eth_api_options.ethapi_trace_cache_duration,
+                eth_log_block_cache: cli.eth_api_options.eth_log_block_cache,
+                eth_statuses_cache: cli.eth_api_options.eth_statuses_cache,
+                max_past_logs: cli.eth_api_options.max_past_logs,
+                tracing_raw_max_memory_usage: cli.eth_api_options.tracing_raw_max_memory_usage,
             };
 
             runner.run_node_until_exit(|config| async move {
